refactor: eventDto 이름 변경 및 위치 이동
