+++ backend/src/main/java/com/back/api/queue/dto/response/WaitingQueueBatchEventResponse.java
@@ -1,15 +1,13 @@
-package com.back.api.queue.dto.event;
+package com.back.api.queue.dto.response;
 
 import java.util.Map;
-
-import com.back.api.queue.dto.response.WaitingQueueResponse;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
 // 대기열 일괄 업데이트 이벤트
 // broadcast 방식으로 모든 대기자에게 한번에 전송
 @Schema(description = "대기열 BroadCast DTO")
-public record WaitingQueueBatchEvent(
+public record WaitingQueueBatchEventResponse(
 
 	@Schema(description = "이벤트 ID", example = "1")
 	Long eventId,
@@ -17,7 +15,7 @@ public record WaitingQueueBatchEvent(
 	@Schema(description = "대기열 사용자별 실시간 상태 업데이트 맵")
 	Map<Long, WaitingQueueResponse> updates
 ) {
-	public static WaitingQueueBatchEvent from(Long eventId, Map<Long, WaitingQueueResponse> updates) {
-		return new WaitingQueueBatchEvent(eventId, updates);
+	public static WaitingQueueBatchEventResponse from(Long eventId, Map<Long, WaitingQueueResponse> updates) {
+		return new WaitingQueueBatchEventResponse(eventId, updates);
 	}
 }
